+++ dbl/__init__.py
@@ -4,7 +4,7 @@
 DBL Python API Wrapper
 ~~~~~~~~~~~~~~~~~~~~~~
 A basic wrapper for the discordbots.org API.
-:copyright: (c) 2019 Francis Taylor & discordbots.org
+:copyright: (c) 2019 Assanali Mukhanov & discordbots.org
 :license: MIT, see LICENSE for more details.
 """
 
@@ -12,7 +12,7 @@ __title__ = 'dblpy'
 __author__ = 'Francis Taylor'
 __license__ = 'MIT'
 __copyright__ = 'Copyright 2019 Assanali Mukhanov'
-__version__ = '0.3.0'
+__version__ = '0.3.1'
 
 from collections import namedtuple
 
@@ -22,4 +22,4 @@ from .http import HTTPClient
 
 VersionInfo = namedtuple('VersionInfo', 'major minor micro releaselevel serial')
 
-version_info = VersionInfo(major=0, minor=3, micro=0, releaselevel='final', serial=0)
+version_info = VersionInfo(major=0, minor=3, micro=1, releaselevel='final', serial=0)